requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
mysql-connector-python
python-dotenv
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
gunicorn
pytest
reportlab
nltk